LPS-104378 Sending apps home
