+++ modules/apps/asset/asset-tags-admin-web/src/main/java/com/liferay/asset/tags/admin/web/internal/depot/application/AssetTagsAdminDepotApplication.java
@@ -12,10 +12,9 @@
  * details.
  */
 
-package com.liferay.depot.web.internal.application.apps;
+package com.liferay.asset.tags.admin.web.internal.depot.application;
 
 import com.liferay.depot.application.DepotApplication;
-import com.liferay.journal.constants.JournalPortletKeys;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -23,11 +22,11 @@ import org.osgi.service.component.annotations.Component;
  * @author Alejandro Tardín
  */
 @Component(immediate = true, service = DepotApplication.class)
-public class JournalDepotApplication implements DepotApplication {
+public class AssetTagsAdminDepotApplication implements DepotApplication {
 
 	@Override
 	public String getPortletId() {
-		return JournalPortletKeys.JOURNAL;
+		return "com_liferay_asset_tags_admin_web_portlet_AssetTagsAdminPortlet";
 	}
 
 }
