+++ modules/apps/document-library/document-library-web/build.gradle
@@ -22,6 +22,7 @@ dependencies {
 	compileOnly project(":apps:asset:asset-taglib")
 	compileOnly project(":apps:bulk:bulk-selection-api")
 	compileOnly project(":apps:comment:comment-taglib")
+	compileOnly project(":apps:depot:depot-api")
 	compileOnly project(":apps:document-library:document-library-api")
 	compileOnly project(":apps:document-library:document-library-repository-authorization-api")
 	compileOnly project(":apps:document-library:document-library-taglib")
